+++ lib/detour_flutter_plugin.dart
@@ -2,6 +2,7 @@ import 'src/models.dart';
 import 'detour_flutter_plugin_platform_interface.dart';
 
 export 'src/models.dart';
+export 'src/detour_service.dart';
 
 class DetourFlutterPlugin {
   Future<void> configure(DetourConfig config) {
@@ -19,28 +20,11 @@ class DetourFlutterPlugin {
   Stream<DetourResult> get linkStream =>
       DetourFlutterPluginPlatform.instance.linkStream;
 
-  Future<void> resetSession({bool allowDeferredRetry = false}) {
-    return DetourFlutterPluginPlatform.instance.resetSession(
-      allowDeferredRetry: allowDeferredRetry,
-    );
-  }
-
-  Future<void> mountAnalytics() {
-    return DetourFlutterPluginPlatform.instance.mountAnalytics();
-  }
-
-  Future<void> unmountAnalytics() {
-    return DetourFlutterPluginPlatform.instance.unmountAnalytics();
-  }
-
   Future<void> logEvent(
-    String eventName, {
+    DetourEventName eventName, {
     Map<String, dynamic>? data,
   }) {
-    return DetourFlutterPluginPlatform.instance.logEvent(
-      eventName,
-      data: data,
-    );
+    return DetourFlutterPluginPlatform.instance.logEvent(eventName, data: data);
   }
 
   Future<void> logRetention(String eventName) {